project/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tcm-management-system"
version = "0.1.0"
description = "Traditional Chinese Medicine Management System based on Django"
requires-python = ">=3.11"
dependencies = [
  "django==5.2",
  "djangorestframework==3.15.2",
  "django-filter==24.3",
  "djangorestframework-simplejwt==5.3.1",
  "django-cors-headers==4.4.0",
  "PyMySQL==1.1.1",
]
